+++ HISTORY.rst
@@ -21,6 +21,7 @@ Release History
   Ideally IA-S3 would always return XML, but that's not the case as of now.
   Try to dump the HTML in the S3 response if unable to parse XML.
 - Fixed bug where ArchiveSession headers weren't being sent in prepared requests.
+- Fixed bug in ``ia upload --size-hint`` where value was an integer, but requests requries it to be a string.
 
 1.1.0 (2016-11-18)
 ++++++++++++++++++